Move IOState implementation to io_state.cpp
