+++ CMakeLists.txt
@@ -135,6 +135,7 @@ include_directories(
 
 set(${PROJECT_NAME}_standalone_sources
   standalone/src/scanner_v2.cpp
+  standalone/src/io_state.cpp
   standalone/src/laserscan.cpp
   standalone/src/data_conversion_layer/monitoring_frame_msg.cpp
   standalone/src/data_conversion_layer/start_request.cpp
@@ -337,6 +338,7 @@ if(CATKIN_ENABLE_TESTING)
 
   catkin_add_gtest(unittest_laserscan
     standalone/test/unit_tests/api/unittest_laserscan.cpp
+    standalone/src/io_state.cpp
     standalone/src/laserscan.cpp
   )
   target_link_libraries(unittest_laserscan
@@ -346,6 +348,7 @@ if(CATKIN_ENABLE_TESTING)
 
   catkin_add_gtest(unittest_pin_state
     standalone/test/unit_tests/api/unittest_pin_state.cpp
+    standalone/src/io_state.cpp
   )
   target_link_libraries(unittest_pin_state
     ${catkin_LIBRARIES}
@@ -354,6 +357,7 @@ if(CATKIN_ENABLE_TESTING)
 
   catkin_add_gtest(unittest_io_state
     standalone/test/unit_tests/api/unittest_io_state.cpp
+    standalone/src/io_state.cpp
   )
   target_link_libraries(unittest_io_state
     ${catkin_LIBRARIES}
@@ -362,6 +366,7 @@ if(CATKIN_ENABLE_TESTING)
 
   catkin_add_gmock(unittest_laserscan_conversions
     standalone/test/unit_tests/data_conversion_layer/unittest_laserscan_conversions.cpp
+    standalone/src/io_state.cpp
     standalone/src/laserscan.cpp
     standalone/src/data_conversion_layer/monitoring_frame_msg.cpp
     standalone/src/data_conversion_layer/diagnostics.cpp
@@ -373,6 +378,7 @@ if(CATKIN_ENABLE_TESTING)
 
   catkin_add_gtest(unittest_laserscan_ros_conversions
     test/unit_tests/unittest_laserscan_ros_conversions.cpp
+    standalone/src/io_state.cpp
     standalone/src/laserscan.cpp
     standalone/src/data_conversion_layer/monitoring_frame_msg.cpp
     standalone/src/data_conversion_layer/diagnostics.cpp
@@ -396,6 +402,7 @@ if(CATKIN_ENABLE_TESTING)
     standalone/src/data_conversion_layer/monitoring_frame_msg.cpp
     standalone/src/data_conversion_layer/diagnostics.cpp
     standalone/src/data_conversion_layer/monitoring_frame_deserialization.cpp
+    standalone/src/io_state.cpp
   )
   target_link_libraries(unittest_monitoring_frame_msg
     ${catkin_LIBRARIES}
@@ -406,6 +413,7 @@ if(CATKIN_ENABLE_TESTING)
     standalone/test/unit_tests/data_conversion_layer/unittest_monitoring_frame_msg_stamped.cpp
     standalone/src/data_conversion_layer/monitoring_frame_msg.cpp
     standalone/src/data_conversion_layer/diagnostics.cpp
+    standalone/src/io_state.cpp
   )
   target_link_libraries(unittest_monitoring_frame_msg_stamped
     ${catkin_LIBRARIES}
@@ -416,6 +424,7 @@ if(CATKIN_ENABLE_TESTING)
     standalone/src/data_conversion_layer/monitoring_frame_msg.cpp
     standalone/src/data_conversion_layer/monitoring_frame_deserialization.cpp
     standalone/src/data_conversion_layer/diagnostics.cpp
+    standalone/src/io_state.cpp
     standalone/test/unit_tests/data_conversion_layer/unittest_monitoring_frame_serialization_deserialization.cpp
     standalone/test/src/data_conversion_layer/monitoring_frame_serialization.cpp
   )
@@ -471,6 +480,7 @@ if(CATKIN_ENABLE_TESTING)
 
   catkin_add_gmock(unittest_io_state_rosconversions
     test/unit_tests/unittest_io_state_rosconversions.cpp
+    standalone/src/io_state.cpp
   )
   target_link_libraries(unittest_io_state_rosconversions
     ${catkin_LIBRARIES}
@@ -506,6 +516,7 @@ if(CATKIN_ENABLE_TESTING)
     standalone/test/src/communication_layer/scanner_mock.cpp
     standalone/test/src/data_conversion_layer/monitoring_frame_serialization.cpp
     standalone/src/scanner_v2.cpp
+    standalone/src/io_state.cpp
     standalone/src/laserscan.cpp
     standalone/src/data_conversion_layer/monitoring_frame_msg.cpp
     standalone/src/data_conversion_layer/monitoring_frame_deserialization.cpp
@@ -523,6 +534,7 @@ if(CATKIN_ENABLE_TESTING)
   add_rostest_gmock(integrationtest_ros_scanner_node
     test/integration_tests/integrationtest_ros_scanner_node.test
     test/integration_tests/integrationtest_ros_scanner_node.cpp
+    standalone/src/io_state.cpp
     standalone/src/laserscan.cpp
     standalone/src/data_conversion_layer/monitoring_frame_msg.cpp
     standalone/src/data_conversion_layer/diagnostics.cpp
@@ -660,14 +672,12 @@ if(CATKIN_ENABLE_TESTING)
 
     catkin_add_gtest(hwtest_scan_compare_standalone
       test/hw_tests/hwtest_scan_compare_standalone.cpp
-      standalone/src/laserscan.cpp
     )
 
   else()
     # always at least build the test to avoid build breaking changes
     catkin_add_executable_with_gtest(hwtest_scan_compare_standalone
                                             test/hw_tests/hwtest_scan_compare_standalone.cpp
-                                            standalone/src/laserscan.cpp
                                             EXCLUDE_FROM_ALL)
     add_dependencies(tests hwtest_scan_compare_standalone)
 
@@ -690,6 +700,7 @@ if(CATKIN_ENABLE_TESTING)
     add_rostest_gtest(hwtest_scan_compare
       test/hw_tests/hwtest_scan_compare.test
       test/hw_tests/hwtest_scan_compare.cpp
+      standalone/src/io_state.cpp
       standalone/src/laserscan.cpp
     )
     add_dependencies(hwtest_scan_compare
@@ -701,6 +712,7 @@ if(CATKIN_ENABLE_TESTING)
     # always at least build the test to avoid build breaking changes
     catkin_add_executable_with_gtest(hwtest_scan_compare
       test/hw_tests/hwtest_scan_compare.cpp
+      standalone/src/io_state.cpp
       standalone/src/laserscan.cpp
       EXCLUDE_FROM_ALL
     )
