+++ standalone/include/psen_scan_v2_standalone/io_state.h
@@ -16,14 +16,9 @@
 #ifndef PSEN_SCAN_V2_STANDALONE_IO_STATE_H
 #define PSEN_SCAN_V2_STANDALONE_IO_STATE_H
 
+#include <ostream>
 #include <string>
-#include <utility>
 #include <vector>
-
-#include <fmt/format.h>
-#include <fmt/ostream.h>
-
-#include "psen_scan_v2_standalone/util/format_range.h"
 
 namespace psen_scan_v2_standalone
 {
@@ -49,43 +44,7 @@ private:
   bool state_;
 };
 
-inline PinState::PinState(uint32_t pin_id, const std::string& name, bool state)
-  : id_(pin_id), name_(name), state_(state)
-{
-}
-
-inline bool PinState::operator==(const PinState& ps) const
-{
-  return id() == ps.id() && name() == ps.name() && state() == ps.state();
-}
-
-inline bool PinState::operator!=(const PinState& ps) const
-{
-  return !operator==(ps);
-}
-
-inline uint32_t PinState::id() const
-{
-  return id_;
-}
-
-inline std::string PinState::name() const
-{
-  return name_;
-}
-
-inline bool PinState::state() const
-{
-  return state_;
-}
-
-// LCOV_EXCL_START
-inline std::ostream& operator<<(std::ostream& os, const PinState& pin_state)
-{
-  return os << fmt::format(
-             "PinState(id = {}, name = {}, state = {})", pin_state.id(), pin_state.name(), pin_state.state());
-}
-// LCOV_EXCL_STOP
+std::ostream& operator<<(std::ostream& os, const PinState& pin_state);
 
 //! @brief Represents the set of all I/Os of the scanner and their states.
 class IOState
@@ -101,27 +60,7 @@ private:
   std::vector<PinState> output_{};
 };
 
-inline IOState::IOState(std::vector<PinState> input, std::vector<PinState> output)
-  : input_(std::move(input)), output_(std::move(output))
-{
-}
-
-inline const std::vector<PinState>& IOState::input() const
-{
-  return input_;
-}
-
-inline const std::vector<PinState>& IOState::output() const
-{
-  return output_;
-}
-
-inline std::ostream& operator<<(std::ostream& os, const IOState& io_state)
-{
-  return os << fmt::format("IOState(input = {}, output = {})",
-                           util::formatRange(io_state.input()),
-                           util::formatRange(io_state.output()));
-}
+std::ostream& operator<<(std::ostream& os, const IOState& io_state);
 }  // namespace psen_scan_v2_standalone
 
 #endif  // PSEN_SCAN_V2_STANDALONE_IO_STATE_H
